rename dbpedia.py
